+++ package.json
@@ -11,5 +11,10 @@
     "next": "12.1.5",
     "react": "18.1.0",
     "react-dom": "18.1.0"
+  },
+  "devDependencies": {
+    "@types/node": "^17.0.30",
+    "@types/react": "^18.0.8",
+    "typescript": "^4.6.4"
   }
 }
